fix(selection) esc key switching
