+++ src/index.html
@@ -645,7 +645,7 @@
 				<header>
 					<h2>Update Available!</h2>
 				</header>
-				<p>A new version of the editor is available. This will clear your local storage. Press update to reload and apply it now, but if you are currently drawing press cancel...</p>
+				<p>A new version of the editor is available.<br/>This will clear your local storage.<br/>Press update to reload and apply it now,<br/>but if you are currently drawing press cancel</p>
 				<h4>AND SAVE BEFORE UPDATING!</h4>
 				<nav>
 					<button id="update-reload" aria-label="Reload and Update">
